Public API should accept a SyntaxNode instead of ExpressionSyntax
