+++ src/Compilers/CSharp/Portable/Binder/Binder_Deconstruct.cs
@@ -468,7 +468,6 @@ namespace Microsoft.CodeAnalysis.CSharp
 
         private BoundTupleLiteral DeconstructionVariablesAsTuple(CSharpSyntaxNode syntax, ArrayBuilder<DeconstructionVariable> variables, DiagnosticBag diagnostics, bool hasErrors)
         {
-            bool trackInferredNames = this.Compilation.LanguageVersion.DisallowInferredTupleElementNames();
             int count = variables.Count;
             var valuesBuilder = ArrayBuilder<BoundExpression>.GetInstance(count);
             var typesBuilder = ArrayBuilder<TypeSymbol>.GetInstance(count);
@@ -498,7 +497,10 @@ namespace Microsoft.CodeAnalysis.CSharp
             uniqueFieldNames.Free();
             ImmutableArray<string> tupleNames = namesBuilder.ToImmutableAndFree();
 
-            var inferredPositions = trackInferredNames ? tupleNames.SelectAsArray(n => n != null) : default(ImmutableArray<bool>);
+            // only track the inferred positions in C# 7
+            var inferredPositions = this.Compilation.LanguageVersion.DisallowInferredTupleElementNames()
+                ? tupleNames.SelectAsArray(n => n != null)
+                : default(ImmutableArray<bool>);
 
             var type = TupleTypeSymbol.Create(syntax.Location,
                 typesBuilder.ToImmutableAndFree(), locationsBuilder.ToImmutableAndFree(),
